(MOT-4107) refactor(harness): decompose integration scenario runner (#534)
